add: 정밀 균열 검출 and 연산자 추가

+++ crackSegmentation/views.py
@@ -77,8 +77,9 @@ def detailInference(request):
         width = userImg.shape[1]
         height = userImg.shape[0]
 
-        # resized_img = resized_img[(height-1344)//2:((height-1344)//2)+1344, (width-1344)//2:((width-1344)//2)+1344]
-        resized_img = cv2.resize(resized_img, (1344, 1344))
+        if width > 1344 & height > 1344:
+            resized_img = resized_img[(height-1344)//2:((height-1344)//2)+1344, (width-1344)//2:((width-1344)//2)+1344]
+        else: resized_img = cv2.resize(resized_img, (1344, 1344))
 
 
         objectWidth = 0
@@ -105,23 +106,23 @@ def detailInference(request):
         row_cnt = objectWidth//448
         col_cnt = objectHeight//448
 
-        for col_i in range(1, row_cnt+1):
-            for row_i in range(1, col_cnt+1):
-                tp1 = ((row_i-1) * 448, row_i*448)
-                tp2 = ((col_i-1) * 448, col_i*448)
-                print("------------------")
-                print(row_cnt + col_cnt)
-                print(tp1)
-                print(tp2)
-                print("------------------")
-                cv2.imwrite(
-                    'media/cropped' + '/cropped_'+str(row_i)+'*'+str(col_i)+'_'+ str(title) + '_' + str(length) + '.jpg',
-                            addZeroPadding[(row_i-1) * 448:row_i*448, (col_i-1)*448:col_i*448])
-
-                cv2.imwrite(
-                    'templates/static/images/cropped' + '/cropped_' + str(row_i) + '*' + str(col_i) + '_' + str(title) + '_' + str(
-                        length) + '.jpg',
-                    addZeroPadding[(row_i - 1) * 448:row_i * 448, (col_i - 1) * 448:col_i * 448])
+        # for col_i in range(1, row_cnt+1):
+        #     for row_i in range(1, col_cnt+1):
+        #         tp1 = ((row_i-1) * 448, row_i*448)
+        #         tp2 = ((col_i-1) * 448, col_i*448)
+        #         print("------------------")
+        #         print(row_cnt + col_cnt)
+        #         print(tp1)
+        #         print(tp2)
+        #         print("------------------")
+        #         cv2.imwrite(
+        #             'media/cropped' + '/cropped_'+str(row_i)+'*'+str(col_i)+'_'+ str(title) + '_' + str(length) + '.jpg',
+        #                     addZeroPadding[(row_i-1) * 448:row_i*448, (col_i-1)*448:col_i*448])
+        #
+        #         cv2.imwrite(
+        #             'templates/static/images/cropped' + '/cropped_' + str(row_i) + '*' + str(col_i) + '_' + str(title) + '_' + str(
+        #                 length) + '.jpg',
+        #             addZeroPadding[(row_i - 1) * 448:row_i * 448, (col_i - 1) * 448:col_i * 448])
 
         resized_img = cv2.resize(resized_img, (1344, 1344))
 
@@ -138,32 +139,41 @@ def detailInference(request):
         rightBot = resized_img[896:1344, 896:1344]
 
         cv2.imwrite('media/resized' + '/resized_leftTop_' + str(title) + '_' + str(length) + '.jpg', leftTop)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_leftTop_' + str(title) + '_' + str(length) + '.jpg', leftTop)
         cv2.imwrite('templates/static/images/resized/' + '/resized_leftTop_' + str(title) + '_' + str(length) + '.jpg',
                     leftTop)
         cv2.imwrite('media/resized' + '/resized_midTop_' + str(title) + '_' + str(length) + '.jpg', midTop)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_midTop_' + str(title) + '_' + str(length) + '.jpg', midTop)
         cv2.imwrite('templates/static/images/resized/' + '/resized_midTop_' + str(title) + '_' + str(length) + '.jpg',
                     midTop)
         cv2.imwrite('media/resized' + '/resized_rightTop_' + str(title) + '_' + str(length) + '.jpg', rightTop)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_rightTop_' + str(title) + '_' + str(length) + '.jpg', rightTop)
         cv2.imwrite('templates/static/images/resized/' + '/resized_rightTop_' + str(title) + '_' + str(length) + '.jpg',
                     rightTop)
 
         cv2.imwrite('media/resized' + '/resized_leftMid_' + str(title) + '_' + str(length) + '.jpg', leftMid)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_leftMid_' + str(title) + '_' + str(length) + '.jpg', leftMid)
         cv2.imwrite('templates/static/images/resized/' + '/resized_leftMid_' + str(title) + '_' + str(length) + '.jpg',
                     leftMid)
         cv2.imwrite('media/resized' + '/resized_midMid_' + str(title) + '_' + str(length) + '.jpg', midMid)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_midMid_' + str(title) + '_' + str(length) + '.jpg', midMid)
         cv2.imwrite('templates/static/images/resized/' + '/resized_midMid_' + str(title) + '_' + str(length) + '.jpg',
                     midMid)
         cv2.imwrite('media/resized' + '/resized_rightMid_' + str(title) + '_' + str(length) + '.jpg', rightMid)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_rightMid_' + str(title) + '_' + str(length) + '.jpg', rightMid)
         cv2.imwrite('templates/static/images/resized/' + '/resized_rightMid_' + str(title) + '_' + str(length) + '.jpg',
                     rightMid)
 
         cv2.imwrite('media/resized' + '/resized_leftBot_' + str(title) + '_' + str(length) + '.jpg', leftBot)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_leftBot_' + str(title) + '_' + str(length) + '.jpg', leftBot)
         cv2.imwrite('templates/static/images/resized/' + '/resized_leftBot_' + str(title) + '_' + str(length) + '.jpg',
                     leftBot)
         cv2.imwrite('media/resized' + '/resized_midBot_' + str(title) + '_' + str(length) + '.jpg', midBot)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_midBot_' + str(title) + '_' + str(length) + '.jpg', midBot)
         cv2.imwrite('templates/static/images/resized/' + '/resized_midBot_' + str(title) + '_' + str(length) + '.jpg',
                     midBot)
         cv2.imwrite('media/resized' + '/resized_rightBot_' + str(title) + '_' + str(length) + '.jpg', rightBot)
+        cv2.imwrite('crack_width_checker/data/org_img' + '/resized_rightBot_' + str(title) + '_' + str(length) + '.jpg', rightBot)
         cv2.imwrite('templates/static/images/resized/' + '/resized_rightBot_' + str(title) + '_' + str(length) + '.jpg',
                     rightBot)
 
@@ -261,7 +271,8 @@ def removeImgs(request):
         media_template_cropped = "templates/static/images/cropped"
         media_template_visualized = "templates/static/images/visualized"
         media_template_origin = "templates/static/images/origin"
-        otsu_data = "crack_width_checker/data"
+        otsu_data_deep_mask = "crack_width_checker/data/deep_mask"
+        otsu_data_org_img = "crack_width_checker/data/org_img"
         otsu_result = "crack_width_checker/results/"
 
         if (os.path.exists(media_imgs)):
@@ -304,8 +315,11 @@ def removeImgs(request):
             for file in os.scandir((media_template_cropped)):
                 os.remove(file.path)
 
-        if (os.path.exists(otsu_data)):
-            for file in os.scandir((otsu_data)):
+        if (os.path.exists(otsu_data_deep_mask)):
+            for file in os.scandir((otsu_data_deep_mask)):
+                os.remove(file.path)
+        if (os.path.exists(otsu_data_org_img)):
+            for file in os.scandir((otsu_data_org_img)):
                 os.remove(file.path)
 
         if (os.path.exists(otsu_result)):
@@ -331,6 +345,8 @@ def removeUnusefulImgs(request):
         for i in range(0, len(usefulImgs)):
             if(os.path.exists("media/resized/" + usefulImgs[i])):
                 os.remove("media/resized/" + usefulImgs[i])
+            if (os.path.exists("crack_width_checker/data/org_img/" + usefulImgs[i])):
+                os.remove("crack_width_checker/data/org_img/" + usefulImgs[i])
 
         return HttpResponse("remove completed")
 '''
@@ -355,35 +371,35 @@ def visionInferenceInfo(request):
         pic_name = request_body['img_name']
         length = request_body['length']
         pic_name = pic_name.split('.')[0]
+        if os.path.exists('crack_width_checker/results/resized_'+pic_name+'/'+str(length)+'000profiling_re_result_summary.txt'):
+            textFile = open(r'crack_width_checker/results/resized_'+pic_name+'/'+str(length)+'000profiling_re_result_summary.txt', "r")
+            first_line = textFile.readline().split(' ')[1]
+            second_line = textFile.readline().split(' ')[1]
+            first_line = re.sub(r'[\n]', '', first_line)
+            second_line = re.sub(r'[\n]', '', second_line)
+            first_line = float(first_line)
+            second_line = float(second_line)
 
-        textFile = open(r'crack_width_checker/results/resized_'+pic_name+'/'+str(length)+'000profiling_re_result_summary.txt', "r")
-        first_line = textFile.readline().split(' ')[1]
-        second_line = textFile.readline().split(' ')[1]
-        first_line = re.sub(r'[\n]', '', first_line)
-        second_line = re.sub(r'[\n]', '', second_line)
-        first_line = float(first_line)
-        second_line = float(second_line)
+            csvFile = open(r'crack_width_checker/results/resized_'+pic_name+'/profiling_re_result_summary.csv', "r")
+            csvReader = csv.reader(csvFile)
 
-        csvFile = open(r'crack_width_checker/results/resized_'+pic_name+'/profiling_re_result_summary.csv', "r")
-        csvReader = csv.reader(csvFile)
+            index = 0
+            all_crack_length = 0.0
+            average_crack_width = 0.0
+            for line in csvReader:
+                if index == 1: all_crack_length = float(line[0])
+                if index == 5: average_crack_width = float(line[0])
+                index += 1
 
-        index = 0
-        all_crack_length = 0.0
-        average_crack_width = 0.0
-        for line in csvReader:
-            if index == 1: all_crack_length = float(line[0])
-            if index == 5: average_crack_width = float(line[0])
-            index += 1
+            crack_info_dict = {
+                'real_max_width': second_line,
+                'all_crack_length' : all_crack_length,
+                'average_crack_width': average_crack_width
+            }
+            shutil.copyfile('crack_width_checker/results/resized_'+pic_name+'/resized_'+pic_name+'_9_full_width_visualization.jpg',
+                            'templates/static/images/analyzed/analyzed_'+pic_name+'.jpg')
 
-        crack_info_dict = {
-            'real_max_width': second_line,
-            'all_crack_length' : all_crack_length,
-            'average_crack_width': average_crack_width
-        }
-        shutil.copyfile('crack_width_checker/results/resized_'+pic_name+'/resized_'+pic_name+'_9_full_width_visualization.jpg',
-                        'templates/static/images/analyzed/analyzed_'+pic_name+'.jpg')
-
-        return JsonResponse(crack_info_dict)
+            return JsonResponse(crack_info_dict)
